- Introduced obs_output to define a subset of history given as observation to the agent - ObsTempl shiftes to /util - added test for ObsTempl - added helper to compare nested lists - adjusted history
