+++ openmodelica_microgrid_gym/util/__init__.py
@@ -1,5 +1,6 @@
 from .fastqueue import Fastqueue
 from .itertools_ import nested_map, fill_params, nested_depth, flatten, flatten_together
+from .obs_template import ObsTempl
 from .randproc import RandProcess
 from .recorder import EmptyHistory, SingleHistory, FullHistory
 from .transforms import abc_to_alpha_beta, normalise_abc, abc_to_dq0_cos_sin, dq0_to_abc_cos_sin, abc_to_dq0, cos_sin, \
@@ -8,4 +9,4 @@ from .transforms import abc_to_alpha_beta, normalise_abc, abc_to_dq0_cos_sin, dq
 __all__ = ['abc_to_alpha_beta', 'normalise_abc', 'abc_to_dq0_cos_sin', 'dq0_to_abc_cos_sin', 'abc_to_dq0',
            'cos_sin', 'dq0_to_abc', 'inst_power', 'inst_reactive', 'inst_rms', 'dq0_to_abc_cos_sin_power_inv',
            'nested_map', 'fill_params', 'nested_depth', 'flatten', 'flatten_together',
-           'EmptyHistory', 'SingleHistory', 'FullHistory', 'Fastqueue', 'RandProcess']
+           'EmptyHistory', 'SingleHistory', 'FullHistory', 'Fastqueue', 'RandProcess', 'ObsTempl']
